# openapi/MSO-LO-3.0-swagger-resolved.yaml
openapi: 3.0.0
info:
  title: 5G-EVE Adaptation Layer API
  description: 5G-EVE Adaptation Layer API spec.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "3.0"
servers:
- url: /api
  description: SwaggerHub API Auto Mocking
...
paths:
  /nfvo:
    get:
      tags:
      - Local NFVO
      summary: Retrieve list of local NFVO.
      description: |
        By passing in the appropriate options, you can search for
        available Local Orchestrators
      operationId: NFVO.fetch
      parameters:
      - name: NFVO parameters
        in: query
        description: nfvo identifier
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/NFVO parameters'
      - name: Pagination parameters
        in: query
        description: The parameters are used to create pagination
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/Pagination parameters'
      responses:
        200:
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NFVO'
        400:
          description: The request is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuthentication: []
  /nfvo/{nfvoId}:
    get:
      tags:
      - Local NFVO
      summary: Read an individual NFVO.
      description: |
        By passing in the appropriate options, you can search for
        available Local Orchestrators
      operationId: NFVO.getById
      parameters:
      - name: nfvoId
        in: path
        description: Identifier of a Nfvo
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: NFVO found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NFVO'
        400:
          description: The request is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized, Invalid credentials (client or user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuthentication: []
...
  /nfvo/{nfvoId}/ns_instances/{nsInstanceId}:
    get:
      tags:
      - NS instances
      summary: Read an individual NS instance resource.
      description: |
        This operation allows to retrieve an existing NS from a specific Nfvo.
      operationId: Nfvo.NS.fetch
      parameters:
      - name: nsInstanceId
        in: path
        description: Identifier of the network service
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nfvoId
        in: path
        description: Identifier of a Nfvo
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: NS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NsInstance'
        400:
          description: The request is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized, Invalid credentials (client or user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuthentication: []
    delete:
      tags:
      - NS instances
      summary: Delete an individual NS instance resource.
      description: |
        Delete an individual NS instance resource.
      operationId: Nfvo.NS.terminate
      parameters:
      - name: nsInstanceId
        in: path
        description: Identifier of the network service
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nfvoId
        in: path
        description: Identifier of a Nfvo
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        204:
          description: |
            204 No Content The NS instance resource and the associated NS identifier were deleted successfully. The response body shall be empty.
        400:
          description: The request is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized, Invalid credentials (client or user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuthentication: []
  /nfvo/{nfvoId}/ns_instance/{nsInstanceId}/instantiate:
    post:
      tags:
      - NS instances
      summary: Instantiate a NS instance.
      description: |
        Instantiate a NS. The precondition is that the NS instance must have
        been created and must be in NOT_INSTANTIATED state. As a result of the
        success of this operation, the NFVO creates a "NS Lifecycle Operation
        Occurrence" resource for the request, and the NS instance state becomes
        INSTANTIATED.
      operationId: Nfvo.NS.instantiate
      parameters:
      - name: nsInstanceId
        in: path
        description: Identifier of the network service
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: nfvoId
        in: path
        description: Identifier of a Nfvo
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        202:
          description: |
            202 ACCEPTED
        400:
          description: The request is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized, Invalid credentials (client or user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuthentication: []
...
components:
  schemas:
    NFVO:
      required:
      - id
      - name
      - site
      - type
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: OSM-Turin
        type:
          type: string
          example: OSM
        site:
          type: string
          example: ITALY_TURIN
        uri:
          type: string
          format: url
          example: https://osm-turin.5g-eve.eu
        createdAt:
          type: string
          format: date-time
          example: 2016-08-29T09:12:33.001Z
        updatedAt:
          type: string
          format: date-time
          example: 2016-08-29T09:12:33.001Z
    NsInstance:
      required:
      - id
      - nsInstanceDescription
      - nsInstanceName
      - nsState
      - nsdId
      type: object
      properties:
        id:
          type: string
          description: Identifier of the NS instance.
          format: uuid
        nsInstanceName:
          type: string
          description: Human readable name of the NS instance.
        nsInstanceDescription:
          type: string
          description: Human readable description of the NS instance.
        nsdId:
          type: string
          description: Identifier of the NSD on which the NS instance is based.
          format: uuid
        nsState:
          $ref: '#/components/schemas/NsStateType'
    NsStateType:
      type: string
      description: "State of the NS instance. Permitted values:\n * `NOT_INSTANTIATED`:\
        \ The NS instance is terminated or not instantiated.\n * `INSTANTIATED`: The\
        \ NS instance is instantiated.\n \n"
      enum:
      - NOT_INSTANTIATED
      - INSTANTIATED
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: human-readable  (es. 'PageNotFound', 'BadArguments')
        message:
          type: string
        target:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorResponse_details'
      description: Error rappresentation
...
    ErrorResponse_details:
      type: object
      properties:
        code:
          type: string
        target:
          type: string
        message:
          type: string
...